Add yet another intel library
There are transitive dependencies between Intel libraries that cannot
be discovered using "otool" on the Mac, and only show up partway
through a run during a failed dynamic load.

libmkl_avx*.lib are like that.  Include them, and hope that this is
the end of the story.

While I've marked this as relevant to 406, it's not really --- this is
more the fault of not actually testing OS X installers when Xyce!31 was
being worked on.

Xyce#406