Modifies OpenMP check to include Trilinos library versions
The OpenMP check previously included a check of both the Epetra
configuration and Kokkos configuration.  The later check of Kokkos
was not valid for Trilinos 12.12, but is valid for more recent
versions of Trilinos.  Thus, it was suggested that the OpenMP check
is written to specifically state which check is valid for which
version of Trilinos.  That way, when Trilinos 12.12 support is
dropped, the OpenMP check can be easily updated.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/182