Removes configure option Xyce_TEST_SOLN_VAR_MAP
The Xyce_TEST_SOLN_VAR_MAP configure option was solely used to enable
or disable the ability to use a netlist file option to output the
namesmap.  This is overridden by the command line option -namesfile.
So, it makes little sense that there is a build option for the
netlist option when the command line option is always available.
Therefore, it must go "bye bye!"