Adds capability for writing out circuit and device graph to file
Over the course of many, many years it has been useful to look at the
circuit and device graph generated by Xyce.  Until now, this required
some hackage.  To facilitate collaboration with users who are studying
network partitioning and using ML for circuit identification, this has
been added as a topology option:

.options topology outputgraph=[0,1,2]

Where 0 is off, by default.  The circuit graph is written to
<netlist_filename>_circuitgraph with option 1 and device graph is
written to <netlist_filename>_devicegraph with option 2.

This is specifically being developed to support the ASC IC project
to use ML on circuit topologies for categorization and identification.