Improve frequency domain source loading
The frequency domain source loading involves frequency comparison
between the source frequency and the current frequency that is evaluated
for HB analysis. In the case that the source frequency is calculated
using expressions, it can have small numerical noise. Some types of
sources can handle this correctly for frequency comparison. But some
types of sources did not handle this correctly and this can lead
to wrong frequency domain evaluation. This commit fixes the frequency
comparsion for SIN voltage source

Xyce#325 Xyce#403