Move error check on AC SENS line to initialization
Some of the error checking for a valid/complete .SENS line in the AC
analysis class were in a function called during the simulation run.
This commit moves those checks to a sensitivity setup function that is
called during initialization. Note that while this change catches the
error earlier in the program execution, it also slightly modifies the
exit behavior. From its previous location, the error check would print
an error message to the user and the return value would propagate all
the way up and trigger the simulation to end. In this case, Xyce still
prints the Solution Summary and statistics information. After moving
the error checks, the return value is actually dropped (an upstream
calling function doesn't check/handle the return value). With this
commit, the error message from the Report::UserError0 call is handled
at the next SafeBarrier call, causing the simulation to end. Exiting
this way, Xyce doesn't print the Solution Summary and statistics.
Instead, there's an abort message after the error message.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/371
Xyce#371