Implement subcircuit multiplier capability
Xyce#267
"Support Multiplier "M=" capability to subcircuits"

This commit implements implicit subcircuit multipliers for subcircuit
instance lines.  Of course, Xyce could already handle multipliers if
they were applied explicitly in the subcircuit.  But this issue was
about applying them implicitly, when the multiplier is specified on the
"X" subcircuit line, but is not explicitly applied to any device instances
inside that subcircuit in the input file.

This is a feature that most simulators support, and is frequently used
in IC designs.

With this commit, this sort of netlist fragment:

Xtest node1 node2 test M=5
.subckt test A B
R1 A B 0.5
.ends

Is equivalent to:

Xtest node1 node2 test
.subckt test A B
R1 A B 0.5 M=5
.ends

If subcircuit devices are already applying a multiplier, the implicit
multiplier will override them.  The one exception is nested subcircuits,
in which case the multipliers get multiplied together and the product of
all the implicit multipliers in the subcircuit hierarchy are used.

To help with debugging, I've added a new command line option, to disable
the implicit multiplier features.  I mainly did this to be able to
quickly test if regression test failures are due to these changes, or
due to something else.

There was one test in our test suite that used
subcircuit multipliers, but since its gold standard was generated with
an older version of Xyce, they had been ignored, and thus the gold
standard was wrong. I've confirmed that the only differences were due to
this new feature.

The new command line argument is:

Xyce -subckt_multiplier <option>  netlist.cir

where <option> can either be "true" or "false".  The default is "true"
and "false" is only intended for debugging purposes.