Add error trap for circular param dependence
Xyce/Xyce#405
"Fix inline subcircuit param dependencies"

GH-60
"Xyce does not resolve parameters referencing other parameters on the
same line"

This issue is mostly fixed, but the original fix did not include an
error trap for circular dependencies.  This commit adds one and makes it
a fatal error.  So now a dependency like this won't be allowed:

Xtest node1 node2 test par3=3 par3='par3+2'

Before, Xyce would basically ignore this, and not fully resolve the
parameter, silently.  That was clearly not the right thing to do.  Fixed
now.