Another bugfix for subcircuit multipliers
When you have this scenario:

Xtest node1 node2 sub1 M=4

.subckt sub1 A B PARAMS: M=1
R1 A B 1K  M={M}
.ends

For implicit multipliers, this leads to circular logic, if one treats
the "Xtest" version of M as a normal parameter.  Essentially what
happens is:

Xtest M=4 overides the sub1 M=1.  Then, R1 M={M} becomes M=4.  But then,
the implicit multiplier is applied afterwards, means that R1 M=4*M=16.
This is incorrect and causes the subckt_b test to fail.

In Hspice, "M" is not allowed to be used as a normal parameter inside
the subcircuit.  So, by default in Hspice, the operations are:

Xtest M=4 is set aside.  It doesn't override sub1 M=1.  This
gives: R1 M={M}=1, and then when the implicit multiplier is
applied, R1=4, which is correct.

This commit fixes this issue, to behave the same.  Note, Hspice
has an option that can override this behavior.  It is:

.OPTION XMULT_IN_EXP=YES

or :

.OPTION M_IN_EXP=YES

which changes this behavior, so that "M" is allowed to be used as a
normal parmaeter.  This is not default behavior, however, and I have not
implemented this option in this commit.