Add soln vector output to outputDAEvectors func
Xyce/Xyce#379

I previously added a new output function in the nonlinear solver,
NonLinearSolver::outputDAEvectors.  This function is optionally invoked
when running with the 2-level Newton method, and will output the final
converged DAE vectors at each time step.

I had neglected to include the solution vector in these outputs, but
that feature was needed, so with this commit I'm adding it.

Now, when this is invoked via:

.options twolevel output_dae_vectors=true

The output will be comprised of the following files at each time step:

daeB_???.txt
daeF_???.txt
daeQ_???.txt
soln_???.txt

where "???" is the time step number.