Include parameter precedence in -hspice-ext all
This is related to an older closed issue:

Issue #223
"Add a new subcircuit separator, "." instead of ":" as part
of -hspice-ext"

When I implemented this issue, I did not include it in the
"Xyce -hspice-ext all" options.  Instead, they are handled by a
different command line option, as it was too complicated to handle
directly under -hspice-ext.

So, as a result, this one aspect of code compatibility had to specified
separately.  That is changed now;  setting "Xyce -hspice-ext all" will
also automatically set "Xyce -redefined_params uselast" as well.