Fix for segfault caused by bad M= logic
Xyce/Xyce#267 - subcircuit multipliers

Xyce/Xyce#426 - Fix segfault related to M= parameter

This specific use case for subcircuit M= was broken and caused a seg fault:

Xtest2 3 0 testA M=15
.subckt testA A B
Xtest A B testB
.ends

Where you have a hierarchy of subcircuits, with M= applied to the top
level, but without M= on a subordinate level.  Bad logic resulted in the
parser attempting to use an invalid iterator.  This is fixed now.