Change dev branch version number
It is no longer appropriate for the dev branch to call itself "7.5".

Years ago, we made the dev branch have a "version" number ridiculously
far in the future (like, 8.0) and only fix it at release time.  For
the last couple of releases I've taken to having it report as the next
version, which impacts nothing but packaging --- when we're a dev
build, we don't even report a number with "Xyce -v", we use the git
describe output.