Make XDM citation generic
XDM gets cited in the reference guide, but the reference was to a
specific numbered entry in circuit.bib meaning we have to update it
manually and hunt down all such citations.

I have added a macro just like our UsersGuide macro so that it now
only needs to be done in one place, and circuit.bib can contain as
many different XDM UG reference entries as someone wants.  We'll just
point to the right one when we need it, and only have to do it once.

Xyce#410