Remove declaration of SetSeed
The SetSeed function in N_DEV_DeviceSupport.C was removed in commit
9cc504411, but I forgot to remove the declaration of that function in
N_DEV_DeviceSupport.h.

This is causing link errors on SOME systems, but not any of our
supported systems.

The declaration should have been removed at the same time that the
definition was, but I screwed up.

Xyce#57