Fix look-and-feel 'blank page' issue
Look and feel hates blank pages in documents without any page numbers
or "this page intentionally left blank" oxymorons.  They don't always
complain, and in fact rarely do.  But when they do, it's annoying.
The SAND report style should take care of that for us, but it
doesn't (I have an issue open with the maintainers of the class).

However, it's easy to fix on our end, and I'm putting the fix into
Xyce.sty so we automagically get it in all our documents.

Xyce#433