Add RGEOMOD to the .MODEL card of the BSIM4
Xyce#438
"Elevate BSIM4 instance parameter RGEOMOD to also be model parameter"

This came up when running a circuit using a PDK.  The PDK was setting
the RGEOMOD parameter in the model card for the BSIM4 model.  However,
the original BSIM4 code from Berkeley (which Xyce implemented) only has
RGEOMOD in the instance, so Xyce was printing a lot of warnings.  It was
also causing the source and drain conductances to be set to bad values,
and have to be reset to 1.0e3.

This commit adds RGEOMOD to the model card, and sets it as a default for
the instance value.  If the instance version of RGEOMOD is set, then it
takes priority.