Add text to README_Xyce
When I added the off parameter to VBIC I never noted this
Xyce-specific difference from the official release in README_Xyce.

The primary purpose of this commit is to test whether markdown
references to issues still works after we migrated to cee-gitlab and
subgroups, or whether we have to specify the whole hierarchy, or even
if specifying the whole hierarchy even works.

Xyce#354
issue #354
Xyce/code/uur-proprietary#354