Fix whitespace error
The primary purpose of this commit is to check more ways of specifying
issues in commit messages.  The prior one had a mistake anyway.

If any one of these actually works, we should see a mention of the
commit in the issue report.  Icing on the cake would be for there to
be a link back to the issue report when the commit is viewed in
gitlab.

So far I have only found the unadorned issue number to work that way,
which makes it difficult to refer to issues in other repos.

It might mean we want our issues to be created at a higher group level
than the repository level, but I dunno yet.

Xyce/code/uur-proprietary/Xyce#354
code/uur-proprietary/Xyce#354
uur-proprietary/Xyce#354