various changes to add outputs to 2-level Newton
This is related to surrogate modeling research.  The goal is to extract
various DAE terms for specific subcircuits.  The path chosen has to do
with the 2-level Newton algorithm, as that is the one method in Xyce
that can maintain a separate subcircuit partition.

This commit adds a feature to request that the 2-level inner solver
output its extracted condutance and port currents, and send that output
to files.  These can be requested from the nestlist of the inner problem
via the last 2 commands in this netlist option:

.options twolevel
+ output_dae_vectors=true
+ output_dae_matrices=true
+ output_reduced_conductances=true
+ output_port_currents=true

The first two (output_dat_vectors and output_dae_matrices) were added
before and are complimentary.