Add DAE outputs that filter out port Vsrcs
This is related to a surrogate modeling project.

This commit adds new outputs to the two-level Newton DAE outputs, which
were added recently.  The previous output options related commit are:

.options twolevel
+ output_dae_vectors=true

Before this commit, the output_dae_vectors option would produce files
for the F,Q and B vectors at each time step.  These files would have
names like:

netlist_daeF_###.txt
netlist_daeQ_###.txt
netlist_daeB_###.txt

where ### is the time step number.

After this commit, another output option has been added:

.options twolevel
+ output_dae_vectors_noport=true

When this is used, the following outputs are toggled:

netlist_delF_###.txt
netlist_delB_###.txt

These delF and delQ outputs contain the same vectors as daeF and daeQ,
except that the contributions of the port Vsrcs have been subtracted
out.  The structure of the vectors remains the same, so the entries
associated with the additional current variable (which for Vsrc's wind
up being a voltage drop equation) are still there, but are effectively
zeroed out.

I did not include the "Q" vector in this filtered output, b/c the port
Vsrcs never have a Q vector contribution.

For many subcircuits, this will result in the "delB" file being nothing
but zeros.  The exception will be when there are independent sources
inside of the subcircuit, which are not associated with ports.

As of this commit, the complete set of two-level DAE output options
are the following:

.options twolevel
+ output_dae_vectors=true
+ output_dae_vectors_noport=true
+ output_dae_matrices=true
+ output_reduced_conductances=true
+ output_port_currents=true

For transient simulations this will produce a lot of output.