Export the interface library directory dependency
When Xyce does a find_package(Trilinos), Trilinos sets
a variable for with the location of the Trilinos libraries.
Normally this is fed into Xyce's use of cmake by setting
a CMAKE_PREFIX_PATH.  But if another project is trying
to link to Xyce then the CMAKE_PREFIX_PATH may not
have a path to Trilinos as well.  The solution is
to set the INTERACE_LIBRARY_DIR for the target
Xyce::XyceLib so that these extra directories can
be found.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/187