Add diamond truncation method for multi-tone HB
The diamond truncation method is a standard frequency selection
technique for multi-tone HB and it is avaiable in most RF simulation
tools. This commit adds this diamond truncation method in Xyce and
it now works with APFT based multi-tone HB methods. It can be enabled
by using option selectharms. This new feature is testd in multiple
ways and is verified to work as expected. The intmodmax is required
for diamond truncation method. The method can be more efficient in
some circuits than box truncation method for multi-tone HB.

Xyce/code/uur-proprietary/Xyce#352