Move unit test code from src folder to test folder
This change is part of an effort to setup a new standard/convention
forwriting unit tests in xyce. The setup calls for separating the
source code from the test code. This commit moves the existing unit
tests(that are co-located with the application code) over to the new
test folder. Going forward, the directory structures under xyce/src and
xyce/test should mirror each other.

Remove unit test references from src folder cmake

This commit removes references to unit testing files/resources from
cmake files for application code.

Add and update cmake file for unit tests

This commit adds cmake files to support building the unit tests
projects under the new test folder (xyce/test). This change makes it so
that all the unit testing framework and support for GoogleTest is done
only once. Previously, each set of unit tests was doing this
independently; those redundant calls were removed.

This commit also fixes the existing unit test cmake files; after moving
the test code to a new directory, there were several broken path
references.

Replace unit test build flag with BUILD_TESTING

This commit changes the cmake flag/variable for including the unit test
folder in the build. CTest already defines a variable for this purpose;
BUILD_TESTING. This commit removes references to Xyce_TEST_SUITE and
replaces them with the standard CTest variable, BUILD_TESTING.

CMake changes to support using GTest

This commit starts to standardize the unit testing framework

Loading GoolgeTest is conditioned on the value of the BUILD_TESTING
flag. The value of BUILD_TESTING is setup to be OFF by default, but can
be enabled by passing it as a variable when invoking cmake. The cmake
files for the individual test applications check that GTest was found
(so the test applications that use GTest should only be built if
BUILD_TESTING is enabled and GTest is found).

Moving line to include CTest

In a review of the CMake changes to support unit testing, the argument
was made that including CTest could be considered part of the build
configuration and setup that happens in confg.cmake. So the
include(CTest) operation was moved into the config.cmake flow. The
other argument for moving this line is that it reduces the number of
things that the top level cmake file is directly responsible for.

It was also determined that setting the default for BUILD_TESTING in the
feature modes file is sufficient (a second ‘set’ command is not needed).

Update readme for expression library tests

We changed one of the CMake variables used to control how the expression
library unit tests are built. This commit updates the expression library
README file, which describes the build options for these tests.

Xyce/backlog/project-backlog#262