Remove "if 0" commented code
This file had two disjoint sets of "if 0" comments, both about two
years old.

The first was added in Jan of 2020 in commit df11d21, a commit that
removed excess processParam calls during sweeps.  The commented-out
code was an old block of very verbose debugging output enabled by
DEBUG_TIME, and the "else" block was a less verbose set of debugging
output.

I have removed the old, commented-out block and left only the "else"
code.

The second set of if 0 comments was from commit 30410ee in October
2020, which fixed expression-based sampling in parallel.  During that
effort, new "pack" and "unpack" functions were created, each of which
contained two complete implementations, one in an "if 0" and the other
in the "else".

Since the "else" has been in use since October 2020, the "if 0" part
is dead code.  I have removed that and left only the "else".  If the
"if 0" part is truly valuable then it can be seen in the git history
with "git log -p 30410ee".

Xyce/backlog/project-backlog#6