Remove if 0 commented block
In commit 6b35153 on 29 Aug 2019, a bunch of PCE-related output was
added.  In the process, and older block of code enabling an
"OUTPUT_ALL_SAMPLES" option was commented out with "if 0" and has
remained that way ever since.  This feature is now unconditionally
disabled by having been commented out.

I am removing the commented-out block.

There are other if 0 comments in this file, but I'm only removing this
one in this commit to isolate the change should anyone wish to recover
this block of code later.

Xyce/backlog/project-backlog#6