Remove commented-out debugging output
In commit 63f1c938 on 29 Aug 2019 a block of code that existed solely
for debugging was commented out (it was the only thing done by that
commit).

That block had first been added a few days earlier (already commented
out) in commit 6dd4499, and it got turned on and off a few dimes
during that week.  It has been commented out ever since.

I am removing this commented-out code.  It lives on in git
history,and one can retrieve it with "git log -p" on this commit's
SHA.  Or one can revert this one commit and get it back, as it is
doing nothing else but removing that one block of commented code.

Xyce/backlog/project-backlog#6