Remove tautological "#if"
In commit 153ac67 on 28 Sep 2015, work was done to clean up some
refactor work of Dave Baur's.  In the process of doing this clean up,
a block of code was added inside a "#if 1"/"#endif" pair, meaning it
was unconditionally compiled from the day it was committed, and has
never been touched since.

I am simply removing the "#if 1" and "#endif" here, since they serve
no purpose.

Technically, this is not "removal of #if 0 comments" but it was
mentioned in the issue as something to do as well.  It is the only
instance of this construct in the AnalysisPKG directory.

Xyce/backlog/project-backlog#6