Remove "if 0" from RxnSet
From the very first commit of the RxnSet device, commit fbbe7bd94d on
14 April 2011, there have been commented-out blocks of code that have
never been compiled in.  I think it is safe to declare them dead and
remove them.

In one case, the commented-out code was the only stuff inside an
"else" (not "#else") block, which seems odd.  But by removing the
commented-out code it is now *obvious* that the else is empty.
Perhaps there should have been an error message in there to handle the
case?  Clearly, though, this else block is never reached.

Xyce/backlog/project-backlog#6