Remove ancient commented out code in N_DEV_Reaction
With commit b4a96e838e (7 April 2011), the "N_DEV_Reaction.C" file was
moved from an OUO CVS repo to the UUR CVS repo.  As such, its history
is really obscured.

But prior to that move, the code had contained an "#if 0" comment
block in the "Reaction::output" function.  This commented-out code
just added extra text about concentration dependencies to the output.
It is not clear that this output function is ever even called
anywhere.

At any rate, the origins of this commented-out block are basically
lost, as the only information about them prior to 7 April 2011 was in
CVS "attic" files, which are difficult to find in git now.  But I did
it.

The commented-out code was moved out of the old header and into the
old code in commit Xyce/code/ouo-exp/Xyce_SandiaModels@60aeeb0
in December 2006, but it was first put in to the header in commit
Xyce/code/ouo-exp/Xyce_SandiaModels@42cd60e in June of 2006.

That is, it has been commented out ever since it appeared 16 years
ago.

It is safe to remove it.

Xyce/backlog/project-backlog#6