Remove "#if 0" comments from N_DEV_Neuron3.h
This file has had a block of code commented out by "#if 0" since the
day it was committed in commit c37f75f in 2009.  git blame
misattributes the line to commit 941b98a in 2013 because that was when Dave
Baur rearranged code, but it was there since the beginning.

I am removing it.

Xyce/backlog/project-backlog#6