Remove "#if 0" comments
N_DEV_Neuron4.C has had a block of "#if 0" commented code since the
day it was first committed in 2009 in commit c37f75f.  It is being
removed, along with an associated "#if 0" block in N_DEV_Neuron4.h
that has been around for the same period.

Xyce/backlog/project-backlog#6