Remove "#if 0" comments left over from development
In commit aa7621a822, the BJT was changed over from using the state
vector to store history related to excess phase to using the store
vector.  That's because the state vector should only be used to store
information that needs to be differentiated with respect to time, and
using the "store" vector is preferred for things that are not
differentiated.  During this development, relics of the old
implementation were left inside "#if 0" blocks instead of being
removed.

There was also debugging code in such a block.

I have left in place two "#if 0" blocks that comment out noise-related
parameters in the sensitivity code.  I left them there because the
only reason that they are commented out is that Xyce does not
currently do sensitivity on noise analysis so they are unneeded.  But
given that Xyce MIGHT someday do that type of sensitivity, I left the
comments in place.

Xyce/backlog/project-backlog#6