Remove commented-out incorrect code from capacitor
In commit 7cd9a27 back in 2019, the matrix for the capacitor in AC
sensitivities was fixed to take into account multiplicity.  For
whatever reason, the older code that was incorrect was left in place
in an "#if 0" comment.

Removing the commented out, incorrect code and leaving just the
correct code.

Xyce/backlog/project-backlog#6