Remove "#if 0" comments from BSIM4
In commit d229c696b7 (2020) a big block of commented-out debugging
information was added, and a block of old limiting-related code
commented out and replaced with a nicer-looking use of several
macros.  This was part investigation of a limiting-related convergence
error that was actually just a one-line fix in the end.

I am removing the large block of comments and the old code that was
replaced.

While I'm at it, I've removed two commented-out blocks of SPICE3F5
code that never would have compiled in Xyce and needn't be preserved.

Xyce/backlog/project-backlog#6