Remove "#if 0" commented-out code from MutIndNonlin2
In 2011, in commit f9676924, a block of state-vector related code was
commented out with "#if 0" and the comment "don't need to do this".

11 years of keeping code we don't need to do in comments is enough.

While I was at it, I also removed a block of commented-out
code (commented with C-style comments) immediately preceding.

This file has a lot of commented-out code with C-style comments, but
I'm not attacking that in this sprint.  I only hit this one because it
was immediately prior to a block of "#if 0" comments I was removing.

Xyce/backlog/project-backlog#6