Unconditionalize "#if 1" block
In commit ad7343d49a, the size-dependent params were templatized in
order to set the stage for doing sensitivities on BSIM3 devices.  When
that happened, the old, non-templated code was preserved in the
"#else" part of an "#if 1" conditional.  That happened in 2018.

I am removing the "#if 1" line and the contents of the "#else"/"#endif"
section.  It's been the only code compiled for a while, so the
contents of the "#if 1" part of the conditional should be the only
part that remains.

Xyce/backlog/project-backlog#6