Remove ancient "#if 1" usage
Back in May of 2004, I added a block of code indicating that SIMBUS
was doing something one way, but I thought it should be another.  The
stuff I thought was right was in an "#if 0" and what SIMBUS did was in
the "#else".

A couple of weeks later, we got it all fixed and my "it should be this
way" was the right thing to do.  The "#if 0" was turned to an "#if 1"
and it has remained that way since 26 May 2004.  It has moved around a
bunch since then (so git blame lies about who is at fault).

That's silly.  There is no reason to keep around the dead code that
was only there to be compatible with SIMBUS, given that SIMBUS doesn't
exist and hasn't for many years (there is a SIMBUS you can find on the
web if you search enough, but it's a completely different SIMBUS
unrelated to the one we worked with).

This isn't removal of a "#if 0" block, but rather removal of "#if 1"
tautological conditionals, but it is in the same spirit
and is one of the things mentioned in the issue report as something
that should be done at the same time.

Xyce/backlog/project-backlog#6