Remove single block of "#if 0" from test harness
This test harness has had a commented-out block of code with "#if 0"
since the day it was created.

This test harness has been in the repo since 2009 and is not part of
any test that I can find.  It appears the whole test harness is dead
code.  The bug report for bug
1574 (https://morannon.sandia.gov/bugzilla/show_bug.cgi?id=1574) was
closed claiming it couldn't be tested because the "XyceAsLibrary" test
didn't exist yet.

Honestly, I wouldn't even have bothered fixing this one if it weren't
for the fact that I'm now overwhelmed by how many files in each
package have to be fixed to get rid of "#if 0" and how much work it is
to dig through history to make sure that the code is really dead and
not some valid artificat of current development.  This is the only
file in *this* directory that needs fixing, so I'm fixing it.

Xyce/backlog/project-backlog#6