Removing unused MATRIX_FAILSAFE ifdef
The MATRIX_FAILSAFE ifdef was added in 2000 by Rob Hoekstra, according
to git blame.  This ifdef was always set to 1, and not modified through
the autoconf or cmake infrastructure.  So, it is fine to remove the
ifdef and consolidate the logic that sets the value controlled by the
ifdef.

Xyce/backlog/project-backlog#7