Remove an "#if 0" from DevicePDEInstance
This block of "#if 0"/"#else"/"#endif" has been in this file since
commit 6231754f in 2003.  But that is only because this file was moved
from the main Xyce device package source directory into its own
TCADModels directory with that commit.

In fact, it was in this file from the day it was created on 23 May
2001, when it was actually created by migrating functions out of the
DiodePDE class and into a base class.  Prior to *that* it was first
introduced in the DiodePDE class in commit 4887e1b on 3 May 2001.

The stuff in the "#else" has been the only stuff used for 21 years.
Removing the stuff in the "#if 0" part.

Xyce/backlog/project-backlog#6