Remove commented-out debugging output
In commit 0ea3b102 in March 2015, two big blocks of debugging output
were commented out with "#if 0".

The debugging output was put in in commit 5d3ea70 in Apr 2014.

I am removing the output altogether, since it's been commented out for
7 years and was only live for one year prior to that.  Presumably the
debugging is done.

If this debugging info should be of interest in the future, this
commit could always be temporarily reverted locally.

Xyce/backlog/project-backlog#6