Remove commented-out declarations
In commit 619b7bd294 in November 2012, some vectors got changed from
type double to type "pdeFadType" (a Sacado type).  The old declaration
was left in place inside an "#if 0".

Removing the old declaration, leaving only the one that's been in use
for 10 years.

I note in passing that the very same commit commented out some lines
of other files with "//" rather than "#if 0" and I have NOT gone
through those other files to clean up those commented-out lines.  That
is an activity for another issue.

Xyce/backlog/project-backlog#6