Remove most of the "#if 0" from PDE_2DMesh
All of the "#if 0" instances in this file except one date back to the
day it was created in commit 6231754f78 in July of 2003.  But that
commit just moved the files from one directory to another.  In fact,
some of them date back as far as 2002.

One of the "#if 0" instances was created in commit 2b5ae00291 by DNS
in August 2005 as part of a clean-up of compiler warnings.

One of the blocks of "#if 0" commented code is very large and in fact
contains TWO nested "if 0" blocks with a nested "#else".  This code
contains comments indicating that it needs to be fixed.  It is so
large and complex that I am uncomfortable touching it even though
technically it should be removed and if it ever needs to be fixed it
should be recovered from git history.  So I am leaving it be.

I have removed all the others.

Xyce/backlog/project-backlog#6