Removes MS_FACTORING2 ifdef and code therein
The N_DEV_MutIndNonLin2 has an ifdef MS_FACTORING2, that was committed
to the code in 2011.  According to Rich, this code is abandoned and
can be removed.  It is not integrated into autoconf or cmake, so enabling
that code could only be done via the CPPFLAGS.  Nothing else needs to be
done other than removing the ifdef and the code contained in those ifdefs.

Xyce/backlog/project-backlog#7