Remove commented-out old code from N_IO_CircuitContext
In commit 4c17fb6a27 in May 2021, in order to improve netlist setup
time and make globally-scoped .params be equivalent to .global_param,
a bunch of code started using "getMutableParam" calls instead of
expression eval calls.

In this file, the old usage was left inside an "#if 0" and the new in
the "#else" block.  No reason to keep it, as it can always be found in
git history if one needs to see the old code.

Xyce/backlog/project-backlog#6