Remove commented-out code in N_IO_CmdParse
In commit bbb909f in Feb 2007, a commit was made to fix broken 2-level
with mpich.  Part of this commit commented out a block of code in
N_IO_CmdParse.

Since this code has been in place now for 15 years without needing the
commented-out stuff, it is safe to delete it.

In fact, the code inside that comment would not compile anymore, as it
references the old style of error reporting function calls.

Xyce/backlog/project-backlog#6