Remove commented out N_IO_Op code
In commit a44ead6c9e in June 2020, .print ops were improved so they
could handle complex numbers properly.  In one such change, the
complex ExpressionOp::get function was changed from declaring a
complex variable and initializing it to the return value of an
expression evaluate call in the real part and hard-coded zero in the
imaginary part (because the old expression library could only handle
real numbers) to declaring a complex variable and initializing it to
zero, then changing its value with an evaluate call that fills in the
whole complex number.

For whatever reason, the old behavior was commented out instead of
removed.

Now the old code is removed.

Xyce/backlog/project-backlog#6