Remove commented-out code N_IO_OutputterLocal
In commit a085538 (June 2020) a block of code was commented out that
had been inserted in commit a44ead6 just the day before.  The commit
log for the former commit states that the fix was not a good idea, and
that there was already a work-around for the issue (which was bugzilla
issue 1032 on Joseki --- an issue that is still open as I write this,
but which will probably get migrated to a new gitlab issue soon).

If this issue needs to be revisited and fixed properly and this block
of code is useful to that process, this commit can be reverted to get
it back.

https://joseki-srn.sandia.gov/bugzilla/show_bug.cgi?id=1032
Xyce/backlog/project-backlog#6