Remove commented-out code from ParameterBlock
In commit caa3311d41 (fixing a bug in vector-composite paramters in
December 2009), a block of code was commented out with an "#if 0" and
the "#else" is nothing but a comment.  I am removing the if and
leaving the comment.

In May 2008 in commit 3f5bc6b10c a commented-out block that was
effectively a no-op was added that contained a loop with no contents
and a comment "Error checks:  (flesh these out later)"

If the error checks need to be fleshed out even later than today, the
block can be reconstructed.  The comment is gone.
Xyce/backlog/project-backlog#6