Remove "#if 1" from 2DPDEInstance
This file contained two instances of "#if 1"/"#else"/"#endif" code.
It has contained these unconditional conditionals since it was
created.  It was created by extracting code out of "N_DEV_2DPDE.C"  in
comiit 8f60100 in July 2003, but the unconditional conditional was
already present in *that* file when it was moved from the top-level
device source directory into the TCADModels directory a few weeks
prior.

Unfortunately, the trail gets cold right around there and it is hard
to trace the history further --- "N_DEV_2DPDE.C" has since been
deleted and it takes more work than it is worth to hunt through the
history to find the original commit that introduced this "conditional"
compilation.  And there is little value to the exercise, because we
know it is prior to July 2003 --- and therefore has been
unconditionally using the "#if 1" part of the construct for 19 years.

Though it is pointeless to do so, I did spend a bunch of time trying
to find the origin point of this conditional, and failed.  I did,
however, find a lot of history of this file and of various "if 1"
blocks coming and going during development (it is a reasonable and
common development process) --- these two simply appear to have been
forgotten in a cleanup step.

I am whacking the conditional construct and leaving only the part that
has been getting compiled in.

Xyce/backlog/project-backlog#6