Remove commented-out code from test harness
Since it was created in 2018 (in commit 96764aa1f8),
twoLevelNewtonLinear.h has had two blocks  of commented-out code.  One
block commented out the entire contents of the interfaceFunctions
function, the other commented out a block of debugging code.

The test harness is used in the test cases in the TwoLevelAPI
directory of the test suite.  The commented-out code is clearly not
needed.

If one needs to resurrect these comments, one can revert this commit,
which does nothing but remove the comments from this one file.

Xyce/backlog/project-backlog#6